+++ 099_eval3/ships.hpp
@@ -13,7 +13,7 @@ class Ship {
   std::string ship_type;
   std::string source;
   std::string dest;
-  unsigned total_capacity;
+  unsigned long total_capacity;
   unsigned used_capacity;
   std::vector<std::string> capabilities;
   std::vector<Cargo> cargos_carried;
@@ -24,7 +24,7 @@ class Ship {
        const std::string & ship_type_,
        const std::string & source_,
        const std::string & dest_,
-       unsigned total_capacity_,
+       unsigned long total_capacity_,
        unsigned used_capacity_,
        const std::vector<std::string> & capabilities_,
        const std::vector<Cargo> & cargos_carried_);
